+++ src/realm/sync/noinst/client_impl_base.cpp
@@ -148,7 +148,6 @@ ClientImpl::ClientImpl(ClientConfig config)
     , m_fast_reconnect_limit{config.fast_reconnect_limit}
     , m_reconnect_backoff_info{config.reconnect_backoff_info}
     , m_disable_upload_activation_delay{config.disable_upload_activation_delay}
-    , m_dry_run{config.dry_run}
     , m_enable_default_port_hack{config.enable_default_port_hack}
     , m_disable_upload_compaction{config.disable_upload_compaction}
     , m_fix_up_object_ids{config.fix_up_object_ids}
@@ -195,11 +194,6 @@ ClientImpl::ClientImpl(ClientConfig config)
 
     if (config.reconnect_mode != ReconnectMode::normal) {
         logger.warn("Testing/debugging feature 'nonnormal reconnect mode' enabled - "
-                    "never do this in production!");
-    }
-
-    if (config.dry_run) {
-        logger.warn("Testing/debugging feature 'dry run' enabled - "
                     "never do this in production!");
     }
 
@@ -1650,15 +1644,13 @@ void Session::activate()
     logger.debug("Activating"); // Throws
 
     bool has_pending_client_reset = false;
-    if (REALM_LIKELY(!get_client().is_dry_run())) {
-        bool file_exists = util::File::exists(get_realm_path());
-        m_performing_client_reset = get_client_reset_config().has_value();
+    bool file_exists = util::File::exists(get_realm_path());
+    m_performing_client_reset = get_client_reset_config().has_value();
 
-        logger.info("client_reset_config = %1, Realm exists = %2 ", m_performing_client_reset, file_exists);
-        if (!m_performing_client_reset) {
-            get_history().get_status(m_last_version_available, m_client_file_ident, m_progress,
-                                     &has_pending_client_reset); // Throws
-        }
+    logger.info("client_reset_config = %1, Realm exists = %2 ", m_performing_client_reset, file_exists);
+    if (!m_performing_client_reset) {
+        get_history().get_status(m_last_version_available, m_client_file_ident, m_progress,
+                                 &has_pending_client_reset); // Throws
     }
     logger.debug("client_file_ident = %1, client_file_ident_salt = %2", m_client_file_ident.ident,
                  m_client_file_ident.salt); // Throws
@@ -1950,10 +1942,6 @@ void Session::send_query_change_message()
     REALM_ASSERT(m_pending_flx_sub_set);
     REALM_ASSERT_3(m_pending_flx_sub_set->query_version, >, m_last_sent_flx_query_version);
 
-    if (REALM_UNLIKELY(get_client().is_dry_run())) {
-        return;
-    }
-
     auto sub_store = get_flx_subscription_store();
     auto latest_sub_set = sub_store->get_by_version(m_pending_flx_sub_set->query_version);
     auto latest_queries = latest_sub_set.to_ext_json();
@@ -1976,9 +1964,6 @@ void Session::send_upload_message()
     REALM_ASSERT_EX(m_state == Active, m_state);
     REALM_ASSERT(m_ident_message_sent);
     REALM_ASSERT(!m_unbind_message_sent);
-
-    if (REALM_UNLIKELY(get_client().is_dry_run()))
-        return;
 
     version_type target_upload_version = m_last_version_available;
     if (m_pending_flx_sub_set) {
@@ -2281,12 +2266,6 @@ Status Session::receive_ident_message(SaltedFileIdent client_file_ident)
     }
 
     m_client_file_ident = client_file_ident;
-
-    if (REALM_UNLIKELY(get_client().is_dry_run())) {
-        // Ready to send the IDENT message
-        ensure_enlisted_to_send(); // Throws
-        return Status::OK();       // Success
-    }
 
     // if a client reset happens, it will take care of setting the file ident
     // and if not, we do it here